👌 IMPROVE: Refactor the codebase
