+++ happy/constants.py
@@ -1,0 +1,6 @@
+from os.path import expanduser
+
+
+HOME = expanduser("~")
+
+flower_list = ["🌼 ", "🍀 ", "🌻 ", "🌺 ", "🌹 ", "🌸 ", "🌷 ", "💐 ", "🏵️  "]
